% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{calcMetagene}
\alias{calcMetagene}
\title{Metagene matrix calculation}
\usage{
calcMetagene(
  bam,
  regionGR = NULL,
  txdb = NULL,
  txList = NULL,
  readLen = NULL,
  relFreq = TRUE,
  nTx = 2000,
  cdsStartUpstream = 50,
  cdsStartDownstream = 50,
  cdsEndUpstream = 50,
  cdsEndDownstream = 50
)
}
\arguments{
\item{bam}{A \code{GAlignments} object of aligned reads.
(Required).}

\item{regionGR}{A \code{GRanges} object of the target regions to calculate metagene. Note
that all regions must have equal width. If \code{regionGR} is set, \code{txdb} will be
ignored. If \code{NULL}, \code{txdb} must be set. (Default: NULL).}

\item{txdb}{A \code{TxDb} object of genome annotation. See \code{GenomicFeatures} package
for more details. If \code{NULL}, \code{regionGR} must be set. (Default: NULL).}

\item{txList}{A character vector of transcript IDs. Note that the transcript IDs set here
should also be found in the \code{txdb}. (Default: NULL).}

\item{readLen}{A vector of read lengths to use (positive). If \code{NULL}, all
lengths will be kept. (Default: NULL).}

\item{relFreq}{A logical variable indicating if transcript wise relative frequency (
normalized by the total read counts in each metagene region) should be returned instead of
raw read counts. Note that if \code{txdb} is set but \code{regionGR} is \code{NULL},
metagene for both CDS start and end regions will be calculated. In this case, if setting
\code{relFreq} to \code{TRUE}, metagene will be normalized by the total read counts in both
CDS start and end region for each transcript. (Default: TRUE).}

\item{nTx}{A numeric variable of the number of transcripts to keep when \code{txdb} is
set but \code{txList} is \code{NULL}. The top \code{nTx} transcripts with the most reads in
coding start and end regions will be kept. (Default: 2000).}

\item{cdsStartUpstream}{A numeric variable indicating the width to use for the upstream
region of CDS start site (not including CDS start site) if \code{txdb} is set but
\code{regionGR} is \code{NULL}. (Default: 50).}

\item{cdsStartDownstream}{A numeric variable indicating the width to use for the downstream
region of CDS start site (including CDS start site) if \code{txdb} is set but
\code{regionGR} is \code{NULL}. (Default: 50).}

\item{cdsEndUpstream}{A numeric variable indicating the width to use for the upstream
region of CDS end site (including CDS end site) if \code{txdb} is set but
\code{regionGR} is \code{NULL}. (Default: 50).}

\item{cdsEndDownstream}{A numeric variable indicating the width to use for the downstream
region of CDS end site (not including CDS end site) if \code{txdb} is set but
\code{regionGR} is \code{NULL}. (Default: 50).}
}
\value{
A \code{list} containing three elements. The first element is metagne, either one
\code{matrix} if \code{regionGR} is set or a list of two \code{matrices} in CDS start and
end regions (with names of start and end). Each row is a transcript, each column is a
position, and a value represents read counts or relative frequency for a transcript at a
position. The second element is the \code{GRanges} for the metagene, either one range if
\code{regionGR} is set or a list of two ranges representing the CDS start and end regions.
This list also contains information of the transcripts selected and flanking sequence
lengths for CDS start and end. The third element is an internal variable indicating if
\code{regionGR} is specified or not (1 means \code{regionGR} and 2 means not specified).
}
\description{
Two modes are provided for metagene matrix calculation. If \code{regionGR} is
specified, metagene will be calculated for the regions given in \code{regionGR}. Otherwise,
\code{txdb} must be specified. If \code{txList} is specified, metagene will be calculated in
the CDS start and end regions for the transcripts specified. If \code{txList} is not
specified, the highest-expressed transcripts will be selected and metagene matrix will be
calculated in the CDS start and end regions. Also, the 5'-end position will be kept for the
reads.
}
