% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.calcMetagene}
\alias{.calcMetagene}
\title{Calculate metagene matrix for fixed length regions}
\usage{
.calcMetagene(bamGR, regionGR)
}
\arguments{
\item{bamGR}{A GRanges object of aligned reads.}

\item{regionGR}{A GRanges object of the target regions to calculate metagene. Note that
all regions must have equal width.}
}
\value{
A matrix of metagene.
}
\description{
Calculate metagene matrix for fixed length regions
}
