% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{.calcMetageneStat}
\alias{.calcMetageneStat}
\title{Calculate metagene column-wise summary statistics}
\usage{
.calcMetageneStat(metagene, metageneFun, mode)
}
\arguments{
\item{metagene}{Either a single matrix or a list of two matrices.}

\item{metageneFun}{A character variable of either 'mean' or 'median' indicating how
metagene should be summarized.}

\item{mode}{A numeric variable of either 1 or 2.}
}
\value{
A data.frame with two columns, 'type' and 'value'. If mode == 1, 'type' column will
be filled with 'Region'. If mode == 2, 'type' column will be filled with 'Start' and 'End'.
'value' column is the column summary statistics for each position.
}
\description{
Calculate metagene column-wise summary statistics
}
