% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{.clacMetageneRegion}
\alias{.clacMetageneRegion}
\title{Calculate metagene region}
\usage{
.clacMetageneRegion(regionObj, mode)
}
\arguments{
\item{regionObj}{Either a GRanges object or a list. Must be output from calcMetagene
function.}

\item{mode}{A numeric variable of either 1 or 2.}
}
\value{
If mode == 1, return a numeric vector representing region indices 1, 2, ..., N where
N is the region width. If mode == 2, return a numeric vector combining two numeric vectors
representing CDS start and end region indices.
}
\description{
Calculate metagene region
}
