% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.filterSeqlevels}
\alias{.filterSeqlevels}
\title{Filter seqlevels}
\usage{
.filterSeqlevels(regionGR, seqInfoKeep)
}
\arguments{
\item{regionGR}{A GRanges object of the target regions to calculate metagene. Note that
all regions must have equal width.}

\item{seqInfoKeep}{A Seqinfo object specifying which seqlevels in regionGR to keep.}
}
\value{

}
\description{
Filter seqlevels
}
