% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.getCDSRegions}
\alias{.getCDSRegions}
\title{Get CDS start and end regions}
\usage{
.getCDSRegions(
  txdb,
  cdsStartUpstream,
  cdsStartDownstream,
  cdsEndUpstream,
  cdsEndDownstream
)
}
\arguments{
\item{txdb}{A TxDb object of genome annotation. See GenomicFeatures package for more
details.}

\item{cdsStartUpstream}{A numeric variable indicating the width to use for the upstream
region of CDS start site (not including CDS start site).}

\item{cdsStartDownstream}{A numeric variable indicating the width to use for the downstream
region of CDS start site (including CDS start site).}

\item{cdsEndUpstream}{A numeric variable indicating the width to use for the upstream
region of CDS end site (including CDS end site).}

\item{cdsEndDownstream}{A numeric variable indicating the width to use for the downstream
region of CDS end site (not including CDS end site).}
}
\value{
a list of 3 elements, CDS start region, CDS end region, and transcript id (names of
the list are "start", "end", and "tx"). CDS start and end regions are both GRanges objects
with a mcol of "tx".
}
\description{
Get CDS start and end regions
}
