% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfscore.R
\name{.getFramePosPct}
\alias{.getFramePosPct}
\title{Calculate the percentage of positive elements for each row in a matrix}
\usage{
.getFramePosPct(counts)
}
\arguments{
\item{counts}{A matrix of read counts with three rows. Rows represent three frames, and
columns represent positions.}
}
\value{
A vector showing the percentage of positive elements for each row.
}
\description{
Calculate the percentage of positive elements for each row in a matrix
}
