% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfscore.R
\name{.getORFScoreSign}
\alias{.getORFScoreSign}
\title{Calculate ORFScore sign}
\usage{
.getORFScoreSign(targetFrame, frame1, frame2, frame3)
}
\arguments{
\item{targetFrame}{A numeric variable of the target frame. This frame is expected to have
larger counts than the other two.}

\item{frame1}{A numeric variable of frame 1 counts.}

\item{frame2}{A numeric variable of frame 2 counts.}

\item{frame3}{A numeric variable of frame 3 counts.}
}
\value{
A numeric variable (-1 or +1), depending on:
\enumerate{
\item targetFrame == 1: if frame1 > frame2 and frame3, then +1, else -1;
}
}
\description{
Calculate ORFScore sign
}
