% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orfscore.R
\name{.getRawORFScore}
\alias{.getRawORFScore}
\title{Calculate raw ORFScore}
\usage{
.getRawORFScore(frameCounts, p)
}
\arguments{
\item{frameCounts}{A numeric vector of length 3 showing the read counts for the three
frames.}

\item{p}{A numeric vector of length 3 indicating the null ditribution for chi-squared test.}
}
\value{
A numeric variable of the test statistic. If the counts for all three frames are
zero, return NA.
}
\description{
Calculate raw ORFScore
}
