% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{.plot_theme}
\alias{.plot_theme}
\title{Plot theme}
\usage{
.plot_theme(gridX = FALSE, gridY = TRUE)
}
\arguments{
\item{gridX}{a logical variable indicating if drawing x-axis grid lines. (Default: FALSE).}

\item{gridY}{a logical variable indicating if drawing y-axis grid lines. (Default: TRUE).}
}
\value{
a ggplot theme.
}
\description{
Plot theme
}
