% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_ops.R
\name{.shift}
\alias{.shift}
\title{Shift the 5'-end of reads of towards downstream position}
\usage{
.shift(bam, readLen, shiftLen)
}
\arguments{
\item{bam}{A GAlignments object of aligned reads of the same length.}

\item{readLen}{The read length to use.}

\item{shiftLen}{The shift size for this read length.}
}
\value{
A GAlignments object of shifted reads.
}
\description{
Shift the 5'-end of reads of towards downstream position
}
