% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene.R
\name{.subsetMetagene}
\alias{.subsetMetagene}
\title{Subset metagene matrices}
\usage{
.subsetMetagene(metagene, nTx)
}
\arguments{
\item{metagene}{A list of two matrices representing read counts in CDS start and end
regions. Each row is a transcript, each column is a position, and value represents read
counts in a transcript at a position.}

\item{nTx}{A numeric variable of the number of transcripts to keep.}
}
\value{
A filtered list of two matrices where the highest expressed transcripts (in coding
start and end regions) are kept.
}
\description{
Keep transcripts with the highest expression in coding start and end regions.
}
