% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{.trimORF}
\alias{.trimORF}
\title{Trim ORF}
\usage{
.trimORF(orfGRL, trimStart, trimEnd)
}
\arguments{
\item{orfGRL}{A GRangesList object of the ORFs.}

\item{trimStart}{A numeric variable indicating how many bases to trim for ORF start.}

\item{trimEnd}{A numeric variable indicating how many bases to trim for ORF end.}
}
\value{
a GRangesList object of trimmed ORFs.
}
\description{
Trim ORF from both ends. If the trimmed length is longer than one CDS, that
whole CDS will be removed and the trimming will continue on the next CDS using the
remaining trimming length. For example, given an ORF and the first CDS is 3bp, if trimStart
is 6bp, then the first CDS will be removed and the second CDS will be trimmed 3bp.
}
