% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_ops.R
\name{loadBam}
\alias{loadBam}
\title{Load a BAM file}
\usage{
loadBam(
  bamFile,
  param = ScanBamParam(flag = scanBamFlag(isUnmappedQuery = FALSE, isSecondaryAlignment
    = FALSE, isNotPassingQualityControls = FALSE), mapqFilter = 0)
)
}
\arguments{
\item{bamFile}{The path of a BAM file. Note that the BAM index file must also exist in the
same folder and have the same prefix. For example, if the BAM file is aln.bam, then the BAM
index file should be aln.bam.bai. (Required).}

\item{param}{A \code{\link[Rsamtools]{ScanBamParam}} object. See \code{Rsamtools} package
for more details. (Default: ScanBamParam(flag=scanBamFlag(isUnmappedQuery=FALSE,
isSecondaryAlignment=FALSE, isNotPassingQualityControls=FALSE), mapqFilter=0)).}
}
\value{
A \code{\link[GenomicAlignments]{GAlignments}} object of the aligned reads.
}
\description{
Load a BAM file. This function is essentially a wrapper of
\code{\link[GenomicAlignments]{readGAlignments}}. Note that paried-end reads
will be treated as single-end.
}
