% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{metagenePlot}
\alias{metagenePlot}
\title{Metagene plot}
\usage{
metagenePlot(
  metagene,
  color = ifelse(metagene$mode == 1, "steelblue", c("red1", "steelblue")),
  metageneFun = c("mean", "median")
)
}
\arguments{
\item{metagene}{A \code{list} containing three elements. The first element is metagne,
either one \code{matrix} if \code{regionGR} is set or a list of two \code{matrices} in CDS
start and end regions (with names of start and end). Each row is a transcript, each column
is a position, and a value represents read counts or relative frequency for a transcript at
a position. The second element is the \code{GRanges} for the metagene, either one range if
\code{regionGR} is set or a list of two ranges representing the CDS start and end regions.
This list also contains information of the transcripts selected and flanking sequence
lengths for CDS start and end. The third element is an internal variable indicating if
\code{regionGR} is specified or not (1 means \code{regionGR} and 2 means not specified).
Check \code{calcMetagene} function to see more details of metagene calculation and output
object. The \code{metagene} must be output from \code{calcMetagene} function. (Required).}

\item{color}{A character variable if \code{metagene$mode == 1}, or a character vector
specifying the colors for CDS start and end regions if \code{metagene$mode == 2}. (Default:
\code{ifelse(metagene$mode == 1, 'steelblue', c('red1', 'steelblue'))}).}

\item{metageneFun}{A character variable indicating how metagene should be summarized. Must
be selected from \code{c('mean', 'median')}. If \code{'mean'}, column means will be taken.
If \code{'median'}, column medians will be taken. (Default: \code{'mean'}).}
}
\value{
A \code{ggplot} object showing the metagene signal per position. If
\code{metagene$mode == 1}, the x-axis is labeled as 1, 2, ..., N where N is the width of the
metagene region. If \code{metagene$mode == 2}, the CDS start region is labeled as -M, -(M-1)
, -(M-2), ..., -1, 0, 1, ..., (N-1) where M and N are values set by \code{cdsStartUpstream}
and \code{cdsStartDownstream} options in \code{calcMetagene} function. Position 0 in CDS
start region represents the first base of start codon; the CDS end region is labeled as
-(P-1), -(P-2), ..., -1, 0, 1, ..., Q where P and Q are values set by \code{cdsEndUpstream}
and \code{cdsEndDownstream}. Position 0 in CDS end region represents the last base of stop
codon.
}
\description{
Plot metagene signal per position. The \code{metagene} must be output from
\code{calcMetagene} function.
}
