% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dist.R
\name{readGenomeDist}
\alias{readGenomeDist}
\title{Genomic feature distribution of aligned reads}
\usage{
readGenomeDist(
  bam,
  txdb,
  category = c("CDS", "UTR", "Intron", "Intergenic"),
  fiveEndOnly = TRUE,
  ignoreStrand = TRUE
)
}
\arguments{
\item{bam}{A \code{GAlignments} or \code{GRanges} object of aligned reads. (Required).}

\item{txdb}{A \code{TxDb} object of genome annotation. See \code{GenomicFeatures} package
for more details. (Required).}

\item{category}{A vector of characters. By default, distribution of reads on CDS, UTR,
intron, and intergenic regions is calculated. Must be selected from
\code{c('CDS', 'UTR', 'Intron', 'Intergenic')}. (Default: c('CDS', 'UTR', 'Intron',
'Intergenic')).}

\item{fiveEndOnly}{A logical variable indicating if only keeping the 5'-ends of reads.
This option can be used to decrease the ambiguity for reads spanning multiple features.
(Default: TRUE).}

\item{ignoreStrand}{A logical variable indicating if ignoring that reads and features must
be on the same strand (Default: TRUE).}
}
\value{
A \code{data.frame} with 3 columns (\code{feature}, \code{count}, and \code{pct})
of genomic feature, number of reads falling into this feature, and percentage of reads
falling into this feature.
}
\description{
Genomic feature distribution of aligned reads. Reads are assigned to a feature
based on the following priority: CDS, UTR, Intron, and Intergenic.
}
