% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{readGenomeDistPlot}
\alias{readGenomeDistPlot}
\title{Genomic feature distribution plot of aligned reads}
\usage{
readGenomeDistPlot(featureDF, type = c("count", "pct"))
}
\arguments{
\item{featureDF}{A \code{data.frame} of reads genomic feature distribution containing three
columns, \code{feature}, \code{count}, and \code{pct}. This dataframe should be output from
\code{readGenomeDist} function. (Required).}

\item{type}{A character variable indicating if the actual read counts for each read length
or the relative percentage should be plotted. Must be selected from \code{c('count',
'pct')}. (Default: \code{'pct'}).}
}
\value{
A \code{ggplot} object showing the reads genomic feature distribution.
}
\description{
Plot reads genomic feature distribution. The \code{featureDF} dataframe
should be output from \code{readGenomeDist} function.
}
