% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dist.R
\name{readLengthDist}
\alias{readLengthDist}
\title{Read length distribution}
\usage{
readLengthDist(bam)
}
\arguments{
\item{bam}{A \code{GAlignments} object of aligned reads. (Required).}
}
\value{
A \code{data.frame} with 3 columns (\code{readLen}, \code{count}, and \code{pct})
of read length, number of reads with this length, and percentage of reads with this length.
}
\description{
Read length distribution.
}
