% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{readLengthDistPlot}
\alias{readLengthDistPlot}
\title{Read length distribution plot}
\usage{
readLengthDistPlot(readLenDF, color = "steelblue", type = c("count", "pct"))
}
\arguments{
\item{readLenDF}{A \code{data.frame} of read length distribution containing three columns,
\code{readLen}, \code{count}, and \code{pct}. This dataframe should be output from
\code{readLengthDist} function. (Required).}

\item{color}{A character variable indicating the color of the read length distribution
bar plot. (Default: \code{'steelblue'}).}

\item{type}{A character variable indicating if the actual read counts for each read length
or the relative percentage should be plotted. Must be selected from \code{c('count',
'pct')}. (Default: \code{'pct'}).}
}
\value{
A \code{ggplot} object showing the read length distribution.
}
\description{
Plot read length distribution as a barplot. The \code{readLenDF} should be
output from \code{readLengthDist} function.
}
