% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bam_ops.R
\name{shiftReads}
\alias{shiftReads}
\title{Shift the 5'-end of reads of towards downstream direction}
\usage{
shiftReads(bam, readLens = NULL, shiftLens = NULL, fiveEndOnly = TRUE)
}
\arguments{
\item{bam}{A \code{GAlignments} object of aligned reads.
(Required).}

\item{readLens}{A vector of read lengths to use (positive). If \code{NULL}, all
lengths will be kept. (Default: NULL).}

\item{shiftLens}{A vector of shift size (non-negative) for each read length.
If \code{NULL}, reads will not be shifted. If both \code{shiftLens} and \code{readLens}
are not \code{NULL}, they must have the same length. If \code{readlens} is \code{NULL}
and \code{shiftLens} is a single value, all reads will be shifted by this length.
(Default: NULL).}

\item{fiveEndOnly}{A logical variable indicating if only keeping the shifted 5'-end.
(Default: TRUE).}
}
\value{
If \code{fiveEndOnly} is \code{TRUE}, return a \code{GRanges} object of shifted
reads. Otherwise, return a \code{GAlignments} object of shifted reads.
}
\description{
Shift the 5'-end of reads of towards downstream direction. Ribosome
profiling reads are known to have P-site offset. Reads on the positive strand
will be shifted to the right, and reads on the negative strand will be shifted
to the left.
}
